Rename config.py → compose_generation.py
